sme ,re cjamges

+++ blockchains/mv_queries/210_LatestNetworkBlockInfo.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS aave_ethereum.LatestNetworkBlockInfo
 (
-    network_id UInt64,
+    network_id UInt8,
     latest_block_number UInt64,
     latest_block_timestamp DateTime64(6),
     network_time_for_new_block UInt64
