+++ blockchains/mv_queries/410_NetworkBlockInfoDictionary.sql
@@ -1,6 +1,6 @@
 CREATE DICTIONARY IF NOT EXISTS aave_ethereum.NetworkBlockInfoDictionary
 (
-    network_id UInt64,
+    network_id UInt8,
     latest_block_number UInt64,
     latest_block_timestamp DateTime64(6),
     network_time_for_new_block UInt64
